--- src/components/NavBarSims4Dos/NavBarSims4Dos.module.scss
@@ -1,70 +1,0 @@
-.navBarSims4Dos {
-  display: flex;
-  flex-direction: row;
-  position: fixed;
-  top: auto;
-  width: 100%;
-  height: 60px;
-  gap: 100px;
-  border: 1px solid rgb(233, 233, 233);
-  background-color: white;
-  opacity: 90%;
-  align-items: center;
-  justify-content: center;
-  font-size: medium;
-  font-weight: 400;
-  transition: top 0.3s ease, opacity 0.3s ease; 
-}
-
-.sticky {
-  position: fixed;
-  top: 0;
-  width: 100%;
-  z-index: 1000;
-  opacity: 100%; 
-}
-
-.estilosNavBarSims4Dos {
-  display: flex;
-  flex-direction: row;
-  gap: 90px;
-}
-
-.estilosNavBarSims4DosDer {
-  display: flex;
-  flex-direction: row;
-
-  justify-content: center;
-  align-items: center;
-  gap: 40px;
-}
-
-ul li {
-  list-style: none;
-}
-
-.logoSims4 {
-  width: 100px;
-  height: auto;
-}
-
-.textoBoton {
-  background-color: transparent;
-  border: none;
-
-  font-size: medium;
-  font-weight: 400;
-}
-
-.textoBotonAzul {
-  background-color: rgb(0, 0, 202);
-  color: azure;
-
-  width: 170px;
-  height: 50px;
-  border: none;
-  border-radius: 8px;
-
-  font-size: medium;
-  font-weight: 400;
-}
